roll back statestore change
